+++ App/Classes/Events.py
@@ -1,0 +1,8 @@
+class DataValidation:
+    pass
+
+class OptionsSelector:
+    pass
+
+class Downloader:
+    pass
