+++ main.py
@@ -3,8 +3,7 @@ from PySide6.QtWidgets import *
 from PySide6.QtCore import *
 from PySide6.QtGui import *
 from App.Pages.Base import Base
-from App.Classes.Utility import LoadCustomFonts, ConfigManager
-from App.Classes.Router import Router
+from App.Classes.Utility import LoadCustomFonts, ConfigManager, Router
 
 def main():
     app = QApplication(sys.argv)
